% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indic.R
\name{calc_indic}
\alias{calc_indic}
\title{Calculate SEEE indicators}
\usage{
calc_indic(indic, version = NULL, file_paths = NULL, data = NULL)
}
\arguments{
\item{indic}{the name of the indicator}

\item{version}{version number of the indicator. If not provided, the latest
is used}

\item{file_paths}{character vector with the paths of the files used to
calculate the indicator. The order must be the ones described in the
documentation of the indicator}

\item{data}{a list of data.frames. Used if file_paths is not provided. The
order must be the ones described in the documentation of the indicator}
}
\value{
a list with two elements:
}
\description{
Calculate SEEE indicators
}
