% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indic.R
\name{get_indic}
\alias{get_indic}
\title{Get the indicators available on SEEE and their version}
\usage{
get_indic(indic = NULL, type = NULL)
}
\arguments{
\item{indic}{optional, a character vector with the names of the indicators
for which one want information. If not provided, will return all the
available indicators.}

\item{type}{optional, a character vector with the type of the indicators
(among "beta", "evaluation" and "diagnostic")}
}
\value{
a data frame with four columns for each indicator version available
  on the SEEE: its name, version, type and the number of required input files
}
\description{
Get the indicators available on SEEE and their version
}
